Remediación F (paso 5): recepcion_qr lecturas → RecepcionQRService (SQL en UI 14→0)
Extrae las 14 consultas SQL restantes embebidas en RecepcionQRWidget hacia
RecepcionQRService (más el INSERT OR IGNORE de contenedores_qr de _generar_qr_contenedor).
El widget ahora delega toda lectura/escritura al servicio; ruta canónica UI→servicio→DB.

- Nuevos métodos de consulta (SQL preservado verbatim, devuelven filas crudas):
  listar_pos_abiertas, listar_contenedores_disponibles, listar_pendientes_recepcion,
  listar_proveedores_activos, historial_qr, detalle_contenedor, timeline_movimientos,
  obtener_trazabilidad, obtener_trazabilidad_con_proveedor, buscar_proveedores,
  buscar_productos, listar_sucursales_activas, historial_recepciones.
- Nuevo registrar_contenedor (INSERT OR IGNORE + commit) para el QR generado.
- Allowlists: recepcion_qr_widget baja de 14→0 en SQL_IN_UI y de 1→0 en
  COMMIT_ROLLBACK_IN_UI; ratchet decreciente lo bloquea en 0.

Sin regresión: suite de arquitectura 35 fallos/224 pasa idéntica al baseline
(los 35 fallos son pre-existentes: settings/menu/permissions/orchestrator).
test_recepcion_qr_service + ratchet + commit-guardrail en verde.
